+++ server/src/test/formatter_operators.test.ts
@@ -201,6 +201,49 @@ suite('Formatter Operators', () => {
         assert.strictEqual(formatAntlers('{{ left where       right }}'), '{{ left where right }}');
     });
 
+    test('it preserves space on switch and inlines expanded groups', () => {
+        const input = `{{#
+    @name h1
+    @desc The typography h1 partial to render an h1 with \`class\`, \`as\`, \`color\` and \`content\` attributes.
+    @param* content The h1 content.
+    @param as The wrapping element. Defaults to \`h1\`.
+    @param color Set a custom text color utility. Falls back to \`text-neutral\`.
+    @param class Add custom CSS classes.
+#}}
+
+<!-- /typography/_h1.antlers.html -->
+<{{ as or 'h1' }}
+    class="{{ switch
+                (
+                    (format == 'eyebrow') => '',
+                    () => 'text-2xl md:text-4xl font-bold leading-tight'
+                )
+            }}
+        {{ color or 'text-neutral' }} {{ class }}">
+    {{ content | nl2br }}
+    </{{ as or 'h1' }}>
+<!-- End: /typography/_h1.antlers.html -->`;
+        const expected = `{{#
+    @name h1
+    @desc The typography h1 partial to render an h1 with \`class\`, \`as\`, \`color\` and \`content\` attributes.
+    @param* content The h1 content.
+    @param as The wrapping element. Defaults to \`h1\`.
+    @param color Set a custom text color utility. Falls back to \`text-neutral\`.
+    @param class Add custom CSS classes.
+#}}
+
+<!-- /typography/_h1.antlers.html -->
+<{{ as or 'h1' }} class="{{ switch(
+                              (format == 'eyebrow') => '',
+                              () => 'text-2xl md:text-4xl font-bold leading-tight') }}
+        {{ color or 'text-neutral' }} {{ class }}">
+    {{ content | nl2br }}
+</{{ as or 'h1' }}>
+<!-- End: /typography/_h1.antlers.html -->`;
+
+        assert.strictEqual(formatAntlers(input), expected);
+    });
+
     test('it emits switch groups', () => {
         const input = `{{ test variable="{switch(
 (size == 'sm') => '(min-width: 768px) 35vw, 90vw',
@@ -208,11 +251,11 @@ suite('Formatter Operators', () => {
 (size == 'lg') => '(min-width: 768px) 75vw, 90vw',
 (size == 'xl') => '90vw'
 )}" }}`;
-        const expected = `{{ test variable="{switch(
-  (size == 'sm') => '(min-width: 768px) 35vw, 90vw',
-  (size == 'md') => '(min-width: 768px) 55vw, 90vw',
-  (size == 'lg') => '(min-width: 768px) 75vw, 90vw',
-  (size == 'xl') => '90vw')}" }}`;
+        const expected = `{{ test variable="{ switch(
+   (size == 'sm') => '(min-width: 768px) 35vw, 90vw',
+   (size == 'md') => '(min-width: 768px) 55vw, 90vw',
+   (size == 'lg') => '(min-width: 768px) 75vw, 90vw',
+   (size == 'xl') => '90vw')}" }}`;
 
         assert.strictEqual(formatAntlers(input), expected);
     });